:iphone: Make menu render better on small screens.
